# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: New Euro Consultants API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

  # ── VISAS ─────────────────────────────────────────────────────────────────
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}/admin:
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── FAQs ──────────────────────────────────────────────────────────────────
...
  /eligibility:
    post:
      operationId: submitEligibility
      tags: [eligibility]
      summary: Submit eligibility check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EligibilityInput"
      responses:
        "201":
          description: Submission result with score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityResult"
    get:
      operationId: listEligibilitySubmissions
      tags: [eligibility]
      summary: List eligibility submissions (admin)
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EligibilitySubmission"
  /eligibility/{id}:
    get:
      operationId: getEligibilitySubmission
      tags: [eligibility]
      summary: Get eligibility submission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilitySubmission"
    patch:
      operationId: updateEligibilityStatus
      tags: [eligibility]
      summary: Update submission status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EligibilityStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilitySubmission"

  # ── LEADS (contact form) ──────────────────────────────────────────────────
...
  /embassy-updates:
    get:
      operationId: listEmbassyUpdates
      tags: [embassy-updates]
      summary: List embassy updates
      responses:
        "200":
          description: List of updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmbassyUpdate"
    post:
      operationId: createEmbassyUpdate
      tags: [embassy-updates]
      summary: Create embassy update (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbassyUpdateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbassyUpdate"
  /embassy-updates/{id}:
    patch:
      operationId: updateEmbassyUpdate
      tags: [embassy-updates]
      summary: Update embassy update (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbassyUpdatePatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbassyUpdate"
    delete:
      operationId: deleteEmbassyUpdate
      tags: [embassy-updates]
      summary: Delete embassy update (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── NEWSLETTER ────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List admin users
      responses:
        "200":
          description: List of admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Create admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/hero:
    get:
      operationId: listHeroSlides
      tags: [admin]
      summary: List hero slides
      responses:
        "200":
          description: Hero slides
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeroSlide"
    post:
      operationId: createHeroSlide
      tags: [admin]
      summary: Create hero slide (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeroSlideInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeroSlide"
  /admin/hero/{id}:
    patch:
      operationId: updateHeroSlide
      tags: [admin]
      summary: Update hero slide (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeroSlideUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeroSlide"
    delete:
      operationId: deleteHeroSlide
      tags: [admin]
      summary: Delete hero slide (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AdminUser:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
    AdminUserInput:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string

    # ── VISAS ───────────────────────────────────────────────────────────────
...
    BlogPost:
      type: object
      required: [id, title, slug, excerpt, content, category, readTime, published]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        readTime:
          type: string
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
        createdAt:
          type: string
    BlogPostInput:
      type: object
      required: [title, slug, excerpt, content, category, readTime]
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        readTime:
          type: string
        imageUrl:
          type: string
        published:
          type: boolean
    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        readTime:
          type: string
        imageUrl:
          type: string
        published:
          type: boolean

    # ── FAQs ────────────────────────────────────────────────────────────────
...
    EligibilityInput:
      type: object
      required: [destinationCountry, visaPurpose, name, phone, email]
      properties:
        destinationCountry:
          type: string
        visaPurpose:
          type: string
        ageRange:
          type: string
        maritalStatus:
          type: string
        employmentStatus:
          type: string
        monthlyIncome:
          type: string
        bankBalance:
          type: string
        propertyOwnership:
          type: boolean
        previousTravel:
          type: array
          items:
            type: string
        previousRefusals:
          type: boolean
        refusalCountries:
          type: array
          items:
            type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
    EligibilityResult:
      type: object
      required: [id, score, band, recommendations]
      properties:
        id:
          type: integer
        score:
          type: integer
        band:
          type: string
        recommendations:
          type: array
          items:
            type: string
    EligibilitySubmission:
      type: object
      required: [id, destinationCountry, visaPurpose, name, phone, email, score, band, status, createdAt]
      properties:
        id:
          type: integer
        destinationCountry:
          type: string
        visaPurpose:
          type: string
        ageRange:
          type: ["string", "null"]
        maritalStatus:
          type: ["string", "null"]
        employmentStatus:
          type: ["string", "null"]
        monthlyIncome:
          type: ["string", "null"]
        bankBalance:
          type: ["string", "null"]
        propertyOwnership:
          type: ["boolean", "null"]
        previousTravel:
          type: array
          items:
            type: string
        previousRefusals:
          type: ["boolean", "null"]
        refusalCountries:
          type: array
          items:
            type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        score:
          type: integer
        band:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    EligibilityStatusUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string

    # ── LEADS ───────────────────────────────────────────────────────────────
...
    EmbassyUpdate:
      type: object
      required: [id, country, flag, headline, summary, publishedAt]
      properties:
        id:
          type: integer
        country:
          type: string
        flag:
          type: string
        headline:
          type: string
        summary:
          type: string
        content:
          type: ["string", "null"]
        publishedAt:
          type: string
        createdAt:
          type: string
    EmbassyUpdateInput:
      type: object
      required: [country, flag, headline, summary]
      properties:
        country:
          type: string
        flag:
          type: string
        headline:
          type: string
        summary:
          type: string
        content:
          type: string
        publishedAt:
          type: string
    EmbassyUpdatePatch:
      type: object
      properties:
        country:
          type: string
        flag:
          type: string
        headline:
          type: string
        summary:
          type: string
        content:
          type: string
        publishedAt:
          type: string

    # ── NEWSLETTER ──────────────────────────────────────────────────────────
...
    HeroSlide:
      type: object
      required: [id, headline, subhead, primaryCta, primaryCtaLink, active]
      properties:
        id:
          type: integer
        headline:
          type: string
        subhead:
          type: string
        primaryCta:
          type: string
        primaryCtaLink:
          type: string
        secondaryCta:
          type: ["string", "null"]
        secondaryCtaLink:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        overlayOpacity:
          type: number
        active:
          type: boolean
        sortOrder:
          type: integer
    HeroSlideInput:
      type: object
      required: [headline, subhead, primaryCta, primaryCtaLink]
      properties:
        headline:
          type: string
        subhead:
          type: string
        primaryCta:
          type: string
        primaryCtaLink:
          type: string
        secondaryCta:
          type: string
        secondaryCtaLink:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
        overlayOpacity:
          type: number
        active:
          type: boolean
        sortOrder:
          type: integer
    HeroSlideUpdate:
      type: object
      properties:
        headline:
          type: string
        subhead:
          type: string
        primaryCta:
          type: string
        primaryCtaLink:
          type: string
        secondaryCta:
          type: string
        secondaryCtaLink:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
        overlayOpacity:
          type: number
        active:
          type: boolean
        sortOrder:
          type: integer