# lib/api-zod/src/generated/types/adminUserInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * New Euro Consultants API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserInput {
  email: string;
  name: string;
  password: string;
}