# lib/api-zod/src/generated/types/eligibilityStatusUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * New Euro Consultants API
 * OpenAPI spec version: 0.1.0
 */

export interface EligibilityStatusUpdate {
  status?: string;
  notes?: string;
}