# lib/api-zod/src/generated/types/heroSlideUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * New Euro Consultants API
 * OpenAPI spec version: 0.1.0
 */

export interface HeroSlideUpdate {
  headline?: string;
  subhead?: string;
  primaryCta?: string;
  primaryCtaLink?: string;
  secondaryCta?: string;
  secondaryCtaLink?: string;
  mediaUrl?: string;
  mediaType?: string;
  overlayOpacity?: number;
  active?: boolean;
  sortOrder?: number;
}